sctp_proto

Struct Endpoint

pub struct Endpoint { /* private fields */ }

The main entry point to the library

This object performs no I/O whatsoever. Instead, it generates a stream of packets to send via poll_transmit, and consumes incoming packets and association-generated events via handle and handle_event.

Implementations

impl Endpoint

pub fn new( config: Arc<EndpointConfig>, server_config: Option<Arc<ServerConfig>>, ) -> Self

Create a new endpoint

Returns Err if the configuration is invalid.

pub fn poll_transmit(&mut self) -> Option<Transmit>

Get the next packet to transmit

pub fn set_server_config(&mut self, server_config: Option<Arc<ServerConfig>>)

Replace the server configuration, affecting new incoming associations only

pub fn handle_event( &mut self, ch: AssociationHandle, event: EndpointEvent, ) -> Option<AssociationEvent>

Process EndpointEvents emitted from related Associations

In turn, processing this event may return a AssociationEvent for the same Association.

pub fn handle( &mut self, now: Instant, remote: SocketAddr, local_ip: Option<IpAddr>, ecn: Option<EcnCodepoint>, data: Bytes, ) -> Option<(AssociationHandle, DatagramEvent)>

Process an incoming UDP datagram

pub fn connect( &mut self, config: ClientConfig, remote: SocketAddr, ) -> Result<(AssociationHandle, Association), ConnectError>

Initiate an Association

pub fn reject_new_associations(&mut self)

Unconditionally reject future incoming associations

pub fn config(&self) -> &EndpointConfig

Access the configuration used by this endpoint